requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pytz>=2024.1
aiohttp-socks>=0.8.0
